pyproject.toml
[tool.poetry]
name = "library-management"
version = "0.1.0"
description = "test task fot EM"
authors = ["Alexander_Kibin <3441@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
faker = "^33.0.0"
orjson = "^3.10.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
